backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.1

# File Handling
aiofiles==23.2.1
Pillow==10.1.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
Pillow==10.1.0
pytesseract==0.3.10
imagehash==4.3.1
scikit-learn==1.3.2
opencv-python==4.8.1.78
PyPDF2==3.0.1

# Image processing
opencv-python==4.8.1.78
